feat: job 패키지 생성
