+++ mashup-batch/src/main/java/kr/mashup/branding/job/scorehistory/ScoreHistoryConfig.java
@@ -1,5 +1,11 @@
-package kr.mashup.branding.signup;
+package kr.mashup.branding.job.scorehistory;
 
+import kr.mashup.branding.infrastructure.pushnoti.PushNotiEventPublisher;
+import kr.mashup.branding.repository.schedule.ScheduleRepository;
+import kr.mashup.branding.service.attendance.AttendanceService;
+import kr.mashup.branding.service.scorehistory.ScoreHistoryService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
@@ -9,49 +15,44 @@ import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import kr.mashup.branding.config.BatchConfig;
-import kr.mashup.branding.service.adminmember.AdminMemberService;
-import lombok.RequiredArgsConstructor;
-
-@ConditionalOnProperty(
-    value = BatchConfig.SPRING_BATCH_JOB_NAMES,
-    havingValue = AdminMemberSignupJobConfig.JOB_NAME
-)
 @Configuration
 @RequiredArgsConstructor
-public class AdminMemberSignupJobConfig {
-    static final String JOB_NAME = "admin-member-sign-up";
+@Slf4j
+public class ScoreHistoryConfig {
+    static final String JOB_NAME = "score-history-create";
     private static final String STEP_NAME = JOB_NAME + "-step";
 
     private final JobBuilderFactory jobBuilderFactory;
     private final JobRepository jobRepository;
     private final StepBuilderFactory stepBuilderFactory;
-    private final AdminMemberService adminMemberService;
+    private final ScheduleRepository scheduleRepository;
+    private final AttendanceService attendanceService;
+    private final ScoreHistoryService scoreHistoryService;
+    private final PushNotiEventPublisher pushNotiEventPublisher;
 
     @Bean
-    public Job adminMemberSignUpJob() {
+    public Job scoreHistoryJob() {
         return jobBuilderFactory.get(JOB_NAME)
             .repository(jobRepository)
-            .start(adminMemberSignUpStep())
+            .start(scoreHistoryStep())
             .build();
     }
 
     @Bean
     @JobScope
-    public Step adminMemberSignUpStep() {
+    public Step scoreHistoryStep() {
         return stepBuilderFactory.get(STEP_NAME)
-            .tasklet(adminMemberSignUpTasklet())
+            .tasklet(scoreHistoryTasklet(scheduleRepository, attendanceService, scoreHistoryService, pushNotiEventPublisher))
             .transactionManager(new ResourcelessTransactionManager())
             .build();
     }
 
     @Bean
     @StepScope
-    public Tasklet adminMemberSignUpTasklet() {
-        return new AdminMemberSignupTasklet(adminMemberService);
+    public Tasklet scoreHistoryTasklet(ScheduleRepository scheduleRepository, AttendanceService attendanceService, ScoreHistoryService scoreHistoryService, PushNotiEventPublisher pushNotiEventPublisher) {
+        return new ScoreHistoryTasklet(scheduleRepository, attendanceService, scoreHistoryService, pushNotiEventPublisher);
     }
 }
