+++ mashup-batch/src/main/java/kr/mashup/branding/job/scorehistory/ScoreHistoryTasklet.java
@@ -1,5 +1,8 @@
-package kr.mashup.branding.scorehistory;
+package kr.mashup.branding.job.scorehistory;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 import kr.mashup.branding.domain.attendance.Attendance;
 import kr.mashup.branding.domain.member.Member;
 import kr.mashup.branding.domain.pushnoti.vo.SeminarAttendanceAppliedVo;
@@ -8,7 +11,6 @@ import kr.mashup.branding.domain.scorehistory.ScoreHistory;
 import kr.mashup.branding.infrastructure.pushnoti.PushNotiEventPublisher;
 import kr.mashup.branding.repository.schedule.ScheduleRepository;
 import kr.mashup.branding.service.attendance.AttendanceService;
-import kr.mashup.branding.service.member.MemberService;
 import kr.mashup.branding.service.scorehistory.ScoreHistoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,34 +20,30 @@ import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
 @Slf4j
 @RequiredArgsConstructor
 public class ScoreHistoryTasklet implements Tasklet {
-	private final ScheduleRepository scheduleRepository;
-	private final AttendanceService attendanceService;
-	private final ScoreHistoryService scoreHistoryService;
-	private final PushNotiEventPublisher pushNotiEventPublisher;
+    private final ScheduleRepository scheduleRepository;
+    private final AttendanceService attendanceService;
+    private final ScoreHistoryService scoreHistoryService;
+    private final PushNotiEventPublisher pushNotiEventPublisher;
 
-	@Override
-	@Transactional
-	public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) {
-		List<Schedule> schedules = scheduleRepository.findAllByIsCounted(false);
+    @Override
+    @Transactional
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) {
+        List<Schedule> schedules = scheduleRepository.findAllByIsCounted(false);
         List<Member> pushNotiTargetMembers = new ArrayList<>();
 
-		schedules.forEach(schedule -> {
-			Map<Member, List<Attendance>> attendanceMap = attendanceService.getByScheduleAndGroupByMember(schedule);
-			attendanceMap.forEach((member, attendances) -> {
-				ScoreHistory scoreHistory = scoreHistoryService.createByAttendances(member, schedule, attendances);
-				scoreHistoryService.save(scoreHistory);
-				pushNotiTargetMembers.add(member);
-			});
-			schedule.changeIsCounted(true);
-		});
+        schedules.forEach(schedule -> {
+            Map<Member, List<Attendance>> attendanceMap = attendanceService.getByScheduleAndGroupByMember(schedule);
+            attendanceMap.forEach((member, attendances) -> {
+                ScoreHistory scoreHistory = scoreHistoryService.createByAttendances(member, schedule, attendances);
+                scoreHistoryService.save(scoreHistory);
+                pushNotiTargetMembers.add(member);
+            });
+            schedule.changeIsCounted(true);
+        });
         pushNotiEventPublisher.publishPushNotiSendEvent(new SeminarAttendanceAppliedVo(pushNotiTargetMembers));
-		return RepeatStatus.FINISHED;
-	}
+        return RepeatStatus.FINISHED;
+    }
 }
